pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfmake"
version = "2.0.0"
description = "CLI toolkit for creating, extracting, and compressing PDFs."
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Yuki Obuchi" }
]

dependencies = [
    "Pillow>=9.4.0",
    "reportlab>=3.6.12",
    "tqdm>=4.64.1",
    "PyMuPDF>=1.22.0",
]

[project.scripts]
pdfmake = "pdfmake.__main__:main"

[tool.setuptools.packages.find]
include = ["pdfmake*"]
